First_agent_template_u1/requirements.txt
markdownify
smolagents
requests
duckduckgo_search
pandas
pytz
pydantic==2.10.6
gradio==5.23.1
